requirements.txt
librosa>=0.10.0
matplotlib>=3.7.0
numpy>=1.23.0
opencv-python>=4.7.0.72
python-dateutil>=2.8.2
requests>=2.28.1
soundfile>=0.12.1
